Move MyNode to List package
